[MLR] - Updated folder structure
