--- src/app/app/components/display-card/display-card.component.css
@@ -1,110 +1,0 @@
-.card-content {
-  min-width: 250px;
-}
-
-/* Main Container */
-.main {
-  background: var(--bg-white);
-  border-radius: 2px;
-  box-shadow: 0px 0px 1.5px var(--bright-blue);
-  animation: fadeIn1 0.3s ease-in-out forwards;
-  max-width: 440px;
-}
-
-@keyframes fadeIn1 {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-/* Header */
-.header {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  justify-content: center;
-  color: var(--bright-blue);
-
-  span {
-    text-align: center;
-    font-size: 13px;
-    font-weight: 600;
-    margin: 45px 0px 0px 0px;
-  }
-
-  h1 {
-    font-size: 36px;
-    font-weight: 700;
-    margin: 40px 0px;
-  }
-}
-
-/* Accordion Styling */
-.dynamic-content {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  justify-content: center;
-}
-
-/* Accordion Items */
-.accordion-item {
-  border: unset;
-  width: 100%;
-  font-size: 13px;
-
-  background-color: var(--bg-items);
-  color: var(--bright-blue);
-  padding: 16px 16px;
-  text-align: center;
-  cursor: pointer;
-  transition: background 0.3s;
-  border: 1px solid var(--border-items);
-}
-
-.accordion-item:hover {
-  background-color: var(--accordion-hover);
-}
-
-/* Disabled state */
-.accordion-item:disabled {
-  border: 1px solid var(--border-items);
-  color: var(--text-disabled);
-  pointer-events: none;
-}
-
-/* Accordion Content */
-.accordion-content {
-  background: var(--bg-white);
-  padding: 10px;
-  border-radius: 6px;
-  margin-top: 4px;
-}
-
-ul {
-  list-style: none;
-  padding: 0;
-}
-
-li {
-  margin-bottom: 8px;
-  font-size: 14px;
-  color: #333;
-}
-
-/* Feedback Text */
-.result {
-  display: block;
-  font-size: 12px;
-  color: var(--text-disabled);
-  background-color: var(--bg-disabled);
-  margin: 35px 0px;
-  cursor: pointer;
-}
-
-.pointer {
-  color: var(--bright-blue) !important;
-  cursor: pointer;
-}
