--- src/app/app/pages/activity/activity.component.css
@@ -1,107 +1,0 @@
-.activity-container {
-  width: 600px;
-  min-height: 353px;
-  color: var(--bright-blue);
-  background-color: var(--bg-white);
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  align-items: stretch;
-  box-shadow: 0px 0px 1px var(--bright-blue);
-
-  animation: fadeIn1 0.5s ease-in-out forwards;
-}
-
-.activity-round-card {
-  width: 600px;
-  min-height: 353px;
-  line-height: 2;
-  color: var(--bright-blue);
-  background-color: var(--bg-white);
-  display: flex;
-  flex-direction: column;
-  justify-content: start;
-  align-items: stretch;
-  box-shadow: 0px 0px 1px var(--bright-blue);
-  transform: translateY(10px);
-  animation: fadeIn2 0.3s ease-in-out forwards;
-}
-
-@keyframes fadeIn1 {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-@keyframes fadeIn2 {
-  from {
-    opacity: 0;
-    transform: translateY(10px);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.activity-multi {
-}
-
-.activity-headers {
-  padding: 0px 40px;
-  display: inline-block;
-}
-.activity-name {
-  font-weight: 600;
-  font-size: 16px;
-  margin: 50px 0px 0px 0px;
-  display: inline-block;
-}
-
-.activity-question-no {
-  margin: 40px 0px 40px 0px;
-  font-size: 38px;
-}
-
-.activity-question-txt {
-  margin: unset;
-  background-color: var(--bg-items);
-  box-shadow: inset 0 0.5px 0 var(--bright-blue),
-    /* Top shadow */ inset 0 -0.5px 0 var(--bright-blue); /* Bottom shadow */
-  padding: 20px 35px;
-  text-wrap: wrap;
-}
-.activity-btns {
-  display: flex;
-  align-items: center;
-  flex-direction: row;
-  justify-content: space-evenly;
-}
-
-.activity-btns button {
-  width: 100%;
-  min-height: 100px;
-  color: var(--bright-blue);
-  background-color: var(--bg-white);
-  border: none;
-  text-align: center;
-  font-weight: bold;
-  padding: unset;
-}
-
-.activity-btns > button {
-  transition: background-color 0.3s ease, color 0.3s ease;
-}
-.activity-btns > button:hover {
-  background-color: var(--bright-blue);
-  color: var(--bg-white);
-  cursor: pointer;
-}
-
-.activity-btns > button:active {
-  color: var(--bright-blue);
-  background-color: var(--bg-white);
-  transition: background-color 0.1s ease, color 0.1s ease; /* Faster effect for click */
-}
