--- src/app/app/pages/home/home.component.css
@@ -1,59 +1,0 @@
-/* .home-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-} */
-
-/* Accordion Styling */
-.accordion {
-  display: flex;
-  flex-direction: column;
-  text-align: center;
-  justify-content: center;
-}
-
-/* Accordion Items */
-.accordion-item {
-  border: unset;
-  width: 100%;
-  font-size: 13px;
-
-  background-color: var(--bg-items);
-  color: var(--bright-blue);
-  padding: 16px 16px;
-  text-align: center;
-  cursor: pointer;
-  transition: background 0.3s;
-  border: 1px solid var(--border-items);
-}
-
-.accordion-item:hover {
-  background-color: var(--accordion-hover);
-}
-
-/* Disabled state */
-.accordion-item:disabled {
-  border: 1px solid var(--border-items);
-  color: var(--text-disabled);
-  pointer-events: none;
-}
-
-/* Accordion Content */
-.accordion-content {
-  background: var(--bg-white);
-  padding: 10px;
-  border-radius: 6px;
-  margin-top: 4px;
-}
-
-ul {
-  list-style: none;
-  padding: 0;
-}
-
-li {
-  margin-bottom: 8px;
-  font-size: 14px;
-  color: #333;
-}
