--- src/app/app/pages/page-not-found/page-not-found.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-page-not-found',
-  standalone: true,
-  imports: [],
-  templateUrl: './page-not-found.component.html',
-  styleUrl: './page-not-found.component.css'
-})
-export class PageNotFoundComponent {
-
-}
