+++ src/app/pages/activity/activity.component.html
@@ -1,0 +1,84 @@
+<div *ngIf="activity">
+  <section
+    *ngIf="!hasMultipleRounds && activity.questions.length > 0"
+    class="activity-container"
+  >
+    <div class="activity-headers">
+      <h2 class="activity-name">{{ activity.activity_name.toUpperCase() }}</h2>
+      <h1 class="activity-question-no">Q{{ currentQuestionIndex + 1 }}.</h1>
+    </div>
+    <p
+      class="activity-question-txt"
+      [innerHTML]="
+        activity.questions[currentQuestionIndex].stimulus | boldAsterisk
+      "
+    ></p>
+
+    <div class="activity-btns">
+      <button
+        (click)="saveAnswer(true, activity.questions[currentQuestionIndex])"
+      >
+        CORRECT
+      </button>
+      <button
+        (click)="saveAnswer(false, activity.questions[currentQuestionIndex])"
+      >
+        INCORRECT
+      </button>
+    </div>
+  </section>
+
+  <div *ngIf="showRoundCard && hasMultipleRounds" class="activity-round-card">
+    <div class="activity-headers">
+      <h1 class="activity-name">
+        {{ activity.activity_name.toUpperCase() }}
+      </h1>
+      <h1 class="activity-question-no">ROUND {{ currentRoundIndex + 1 }}</h1>
+    </div>
+  </div>
+
+  <div *ngIf="hasMultipleRounds && !showRoundCard" class="activity-container">
+    <div class="activity-headers">
+      <h1 class="activity-name">
+        {{ activity.activity_name.toUpperCase() }} /
+        {{ activity.questions[currentRoundIndex].round_title.toUpperCase() }}
+      </h1>
+      <h1 class="activity-question-no">Q{{ currentQuestionIndex + 1 }}.</h1>
+    </div>
+
+    <p
+      class="activity-question-txt"
+      [innerHTML]="
+        activity.questions[currentRoundIndex].questions[currentQuestionIndex]
+          .stimulus | boldAsterisk
+      "
+    ></p>
+
+    <div class="activity-btns">
+      <button
+        (click)="
+          saveRoundAnswer(
+            true,
+            activity.questions[currentRoundIndex].questions[
+              currentQuestionIndex
+            ]
+          )
+        "
+      >
+        CORRECT
+      </button>
+      <button
+        (click)="
+          saveRoundAnswer(
+            false,
+            activity.questions[currentRoundIndex].questions[
+              currentQuestionIndex
+            ]
+          )
+        "
+      >
+        INCORRECT
+      </button>
+    </div>
+  </div>
+</div>
