+++ src/app/pages/page-not-found/page-not-found.component.html
@@ -1,0 +1,6 @@
+<div>
+  <h1>Page Not Found</h1>
+  <a href="/">
+    <p>HOME</p>
+  </a>
+</div>
